+++ tests/expectedSchemas/stationsGetItemSchema.js
@@ -207,6 +207,7 @@ module.exports = {
           'nullable': false,
         },
       },
+      'additionalProperties': true,
     },
   },
 }